Another commit, fixed bugs.

+++ Headers/Sudoku.h
@@ -15,6 +15,6 @@ public:
    std::string toString() const;
    std::istream& inputPuzzle(std::istream& input);
 
-   int sudoku[9][9] = { { 0 }, { 0 }}; //This object's puzzle.
+   short int sudoku[9][9] = { { 0 }, { 0 }}; //This object's puzzle.
    int fitness = 1874919423; //This object's fitness, defaults to what should be the max of int.
 };
